python/pytorch/requirements.txt
tqdm
torchvision
numpy
matplotlib
pytorch